+++ lib/logstash/outputs/google_cloud_storage.rb
@@ -62,6 +62,7 @@ require "zlib"
 #      date_pattern => "%Y-%m-%dT%H:00"                          (optional)
 #      flush_interval_secs => 2                                  (optional)
 #      gzip => false                                             (optional)
+#      gzip_content_encoding => false                            (optional)
 #      uploader_interval_secs => 60                              (optional)
 #      upload_synchronous => false                               (optional)
 #    }
@@ -113,8 +114,14 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
   # on every message.
   config :flush_interval_secs, :validate => :number, :default => 2
 
-  # Gzip output stream when writing events to log files.
+  # Gzip output stream when writing events to log files, set
+  # `Content-Type` to `application/gzip` instead of `text/plain`, and
+  # use file suffix `.log.gz` instead of `.log`.
   config :gzip, :validate => :boolean, :default => false
+
+  # Gzip output stream when writing events to log files and set
+  # `Content-Encoding` to `gzip`.
+  config :gzip_content_encoding, :validate => :boolean, :default => false
 
   # Uploader interval when uploading new files to GCS. Adjust time based
   # on your time pattern (for example, for hourly files, this interval can be
@@ -147,13 +154,6 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
 
   config :max_concurrent_uploads, :validate  => :number, :default => 5
 
-  # The path to the service account's JSON credentials file.
-  # Application Default Credentials (ADC) are used if the path is blank.
-  # See: https://cloud.google.com/docs/authentication/production
-  #
-  # You must run on GCP for ADC to work.
-  config :json_key_file, :validate => :string, :default => ""
-
   public
   def register
     @logger.debug('Registering Google Cloud Storage plugin')
@@ -168,6 +168,7 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
     start_uploader
 
     @content_type = @gzip ? 'application/gzip' : 'text/plain'
+    @content_encoding = @gzip_content_encoding ? 'gzip' : 'identity'
   end
 
   # Method called for each log event. It writes the event to the current output
@@ -216,13 +217,13 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
 
   def initialize_path_factory
     @path_factory = LogStash::Outputs::Gcs::PathFactoryBuilder.build do |builder|
-      builder.set_directory @temp_directory
-      builder.set_prefix @log_file_prefix
-      builder.set_include_host @include_hostname
-      builder.set_date_pattern @date_pattern
+      builder.set_directory(@temp_directory)
+      builder.set_prefix(@log_file_prefix)
+      builder.set_include_host(@include_hostname)
+      builder.set_date_pattern(@date_pattern)
       builder.set_include_part(@max_file_size_kbytes > 0)
-      builder.set_include_uuid @include_uuid
-      builder.set_is_gzipped @gzip
+      builder.set_include_uuid(@include_uuid)
+      builder.set_is_gzipped(@gzip)
     end
   end
 
@@ -238,13 +239,13 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
   ##
   # Initializes Google Client instantiating client and authorizing access.
   def initialize_google_client
-    @client = LogStash::Outputs::Gcs::Client.new @bucket, @json_key_file, @logger
+    @client = LogStash::Outputs::Gcs::Client.new(@bucket, @json_key_file, @logger)
   end
 
   ##
   # Uploads a local file to the configured bucket.
   def upload_object(filename)
-    @client.upload_object filename
+    @client.upload_object(filename, @content_encoding, @content_type)
   end
 
   def upload_and_delete(filename)
@@ -262,7 +263,7 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
 
   def initialize_log_rotater
     max_file_size = @max_file_size_kbytes * 1024
-    @log_rotater = LogStash::Outputs::Gcs::LogRotate.new(@path_factory, max_file_size, @gzip, @flush_interval_secs)
+    @log_rotater = LogStash::Outputs::Gcs::LogRotate.new(@path_factory, max_file_size, @gzip, @flush_interval_secs, @gzip_content_encoding)
 
     @log_rotater.on_rotate do |filename|
       @logger.info("Rotated out file: #{filename}")