+++ spec/outputs/google_cloud_storage_spec.rb
@@ -17,5 +17,7 @@ describe LogStash::Outputs::GoogleCloudStorage do
 
   it "should register without errors" do
     expect { subject.register }.to_not raise_error
+
+    subject.close
   end
 end